refactor(controller): SpringApplication과 컨트롤러 분리
